Upload modifica alle classi in chess_model
Eliminato l’uso del byte e abbellito il codice

+++ chess_model/ChessboardModel.java
@@ -5,7 +5,7 @@ public class ChessboardModel {
 	
 	private Piece scacchiera[][] = new Piece[8][8];
 	
-	private ChessboardModel(){
+	private ChessboardModel() {
 		//schiero la prima squadra
 		scacchiera[0][0] = new Rook(1);
 		scacchiera[1][0] = new Knight(1);
@@ -15,16 +15,16 @@ public class ChessboardModel {
 		scacchiera[5][0] = new Bishop(1);
 		scacchiera[6][0] = new Knight(1);
 		scacchiera[7][0] = new Rook(1);
-		for (int i=0; i < 8; i++)
+		for (int i = 0; i < 8; i++)
 			scacchiera[i][1] = new Pawn(1);
 		
 		//riempio la parte centrale della scacchiera di vuoti
-		for(int i = 0;i < 4;i++)
-			for(int j = 0; j <= 7; j++)
+		for (int i = 0;i < 4; i++)
+			for (int j = 0; j <= 7; j++)
 				scacchiera[j][i + 2] = null;
 		
 		//schiero la seconda squadra
-		for (int i=0; i < 8; i++)
+		for (int i = 0; i < 8; i++)
 			scacchiera[i][6] = new Pawn(2);
 		scacchiera[0][7] = new Rook(2);
 		scacchiera[1][7] = new Knight(2);
@@ -36,39 +36,51 @@ public class ChessboardModel {
 		scacchiera[7][7] = new Rook(2);
 	}
 	
-	public static Piece getPezzoInPosizione(byte x, byte y){
+	public static Piece getPezzoInPosizione(int x, int y) {
 		return scacchieraObj.scacchiera[x][y];
 	}
 	
-	public static boolean setPezzoInPosizione(Piece pezzo, byte x, byte y){
+	public static boolean setPezzoInPosizione(Piece pezzo, int x, int y) {
 		scacchieraObj.scacchiera[x][y] = pezzo;
 		return true;
 	}
 	
-	public static void nuovaPartita(){
+	public static void nuovaPartita() {
 		//nessun altra classe può fare una new ChessboardModel
 		//se viene richiamata una nuova partita per la seconda volta occorre fare pulizia...
 		scacchieraObj = new ChessboardModel();
 	}
 	
 	@Override
-	public String toString(){
+	public String toString() {
 		String chessboardToPlay = "";
 		String charPiece = "";
 		chessboardToPlay += "\n   A B C D E F G H\n\n";
-		for(int y = 7; y >= 0; y--){
+		
+		for (int y = 7; y >= 0; y--) {
 			chessboardToPlay += (y + 1) + "  ";
-			for(int x = 0; x < 8; x++){
-				if(scacchiera[x][y] instanceof Rook) charPiece = "R ";
-				else if(scacchiera[x][y] instanceof Bishop) charPiece = "B ";
-				else if(scacchiera[x][y] instanceof King) charPiece = "K ";
-				else if(scacchiera[x][y] instanceof Queen)  charPiece = "Q ";
-				else if(scacchiera[x][y] instanceof Knight) charPiece = "C ";
-				else if(scacchiera[x][y] instanceof Pawn) charPiece = "P ";
+			for (int x = 0; x < 8; x++) {
+				if (scacchiera[x][y] instanceof Rook)
+					charPiece = "R ";
+				else if (scacchiera[x][y] instanceof Bishop)
+					charPiece = "B ";
+				else if (scacchiera[x][y] instanceof King)
+					charPiece = "K ";
+				else if (scacchiera[x][y] instanceof Queen)
+					charPiece = "Q ";
+				else if (scacchiera[x][y] instanceof Knight)
+					charPiece = "C ";
+				else if (scacchiera[x][y] instanceof Pawn)
+					charPiece = "P ";
 				else charPiece = "- ";
-				if(scacchiera[x][y] != null && scacchiera[x][y].team == Team.Team1) charPiece = charPiece.toLowerCase();
+				
+				if (scacchiera[x][y] != null && scacchiera[x][y].team == Team.Team1)
+					charPiece = charPiece.toLowerCase();
+				
 				chessboardToPlay += charPiece;
-				if(x == 7) chessboardToPlay += " " + (y + 1) + "\n";
+				
+				if (x == 7)
+					chessboardToPlay += " " + (y + 1) + "\n";
 			}
 		}
 		chessboardToPlay += "\n   A B C D E F G H";
